+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/repositories/KeyTermRepository.java
@@ -1,0 +1,7 @@
+package ru.dpopkov.knowthenics.repositories;
+
+import org.springframework.data.repository.CrudRepository;
+import ru.dpopkov.knowthenics.model.KeyTerm;
+
+public interface KeyTermRepository extends CrudRepository<KeyTerm, Long> {
+}