+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/repositories/QCollectionRepository.java
@@ -1,0 +1,7 @@
+package ru.dpopkov.knowthenics.repositories;
+
+import org.springframework.data.repository.CrudRepository;
+import ru.dpopkov.knowthenics.model.QCollection;
+
+public interface QCollectionRepository extends CrudRepository<QCollection, Long> {
+}