+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/repositories/QuestionRepository.java
@@ -1,0 +1,7 @@
+package ru.dpopkov.knowthenics.repositories;
+
+import org.springframework.data.repository.CrudRepository;
+import ru.dpopkov.knowthenics.model.Question;
+
+public interface QuestionRepository extends CrudRepository<Question, Long> {
+}