+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/repositories/SourceRepository.java
@@ -1,0 +1,7 @@
+package ru.dpopkov.knowthenics.repositories;
+
+import org.springframework.data.repository.CrudRepository;
+import ru.dpopkov.knowthenics.model.Source;
+
+public interface SourceRepository extends CrudRepository<Source, Long> {
+}