+++ History.md
@@ -2,6 +2,7 @@
 
 #### 2020-12-24
 * Convert Entities to JPA Entities. Closes #22.
+* Add Spring Data JPA Repositories. Closes #23.
 
 #### 2020-12-18
 * Apply master layout to Index Page. Closes #19.